src/expose-yfinance/pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expose-yfinance"
version = "0.1.0"
description = "Expose YFinance live data microservice"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
dependencies = [
  "grpcio==1.74.0",
  "grpcio-tools==1.74.0",
  "yfinance==0.2.65"
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
  "ruff==0.12.11"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 180
target-version = "py311"
extend-exclude = ["src/expose_yfinance/_generated"]
